+++ R/egib_layers.R
@@ -1,0 +1,14 @@
+#' @name egib_layers
+#' @title Available URLs and layers of Land and Building Register (EGiB)
+#'
+#' @description The data frame contains the names of counties,
+#' their identifiers (TERYT, 4 characters), the URL to the EGiB data (as Web
+#' Feature Service) and the names of available layers.
+#'
+#' @importFrom utils data
+#'
+#' @docType data
+#' @keywords dataset EGiB
+#' @examples
+#' egib_layers
+"egib_layers"